커뮤니티-수강후기,질의응답 완료

+++ mentoring/src/main/java/com/karrier/mentoring/entity/Program.java
@@ -59,7 +59,7 @@ public class Program {
 
     private Integer likeCount;
 
-    private Integer averageStar;
+    private Float averageStar;
 
     @ElementCollection
     private List<String> curriculumTitleList;
